change ForumTopic id to `Long`

+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/types/forum/ForumTopic.kt
@@ -1,7 +1,7 @@
 package eu.vendeli.tgbot.types.forum
 
 data class ForumTopic(
-    val messageThreadId: Int,
+    val messageThreadId: Long,
     val name: String,
     val iconColor: IconColor,
     val iconCustomEmojiId: String? = null,
